# use Copy instead of ReadAll in pebble_cache.Get()